Also fix hasMemoryAddress with MemorySegment for Unpooled direct buffers and buffers from the adaptive allocator
