+++ buffer/src/test/java/io/netty/buffer/PooledByteBufAllocatorTest.java
@@ -726,21 +726,6 @@ public class PooledByteBufAllocatorTest extends AbstractByteBufAllocatorTest<Poo
         assertTrue(beforeFreeBytes < afterFreeBytes);
     }
 
-    @Test
-    void directBuffersMustHaveMemoryAddress() throws Exception {
-        // The memory address must always be available when we either have Unsafe available,
-        // or when we have memory segments available (though CleanerJava24 only enables for Java 24+).
-        assumeTrue(PlatformDependent.hasUnsafe() || PlatformDependent.javaVersion() >= 24);
-        PooledByteBufAllocator allocator = newAllocator(true);
-        ByteBuf buf = allocator.directBuffer();
-        try {
-            assertTrue(buf.hasMemoryAddress());
-            assertNotEquals(0L, buf.memoryAddress());
-        } finally {
-            buf.release();
-        }
-    }
-
     @Override
     @Test
     public void testUsedDirectMemory() {
